voice-bridge/server/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
aiohttp>=3.8.0
openai>=1.0.0
openai-whisper>=20231117
python-multipart>=0.0.6
